+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 ## mm/dd/2020
 
 1. [](#new)
-    * Renamed `PageCollectionInterface::nonModular()` into `PageCollectionInterface::pages()` and deprecate the old method
-    * Renamed `PageCollectionInterface::modular()` into `PageCollectionInterface::modules()` and deprecate the old method
+    * Renamed `PageCollectionInterface::nonModular()` into `PageCollectionInterface::pages()` and deprecated the old method
+    * Renamed `PageCollectionInterface::modular()` into `PageCollectionInterface::modules()` and deprecated the old method
 1. [](#bugfix)
     * *Menu Visibility Requires Access* Security option setting wrong frontmatter [login#265](https://github.com/getgrav/grav-plugin-login/issues/265)
     * Accessing page with unsupported file extension (jpg, pdf, xsl) will use wrong mime type [#3031](https://github.com/getgrav/grav/issues/3031)